add clippy-json target
Add a clippy-json target to the ninja script.

This is useful for rust-analyzer, which requires an "external check
command" to function when not using cargo.

This also changes the --emit option in the clippy script to "emit",
without this, clippy will fail with errors about missing
libraries. These would normally be created with the normal compilation
target, but we can't rely on that, because if it fails, clippy-json
never runs, and you won't get error messages to the LSP.